// mobile-app/src/generated/astraflow-api/types.gen.ts
export type AstraflowV1Account = {
    id?: string;
    provider?: string;
    subject?: string;
    email?: string;
    displayName?: string;
    tenantId?: string;
    status?: string;
    createdAt?: string;
    updatedAt?: string;
};
// ...
export type AstraflowV1NativeOAuthTokens = {
    accessToken?: string;
    refreshToken?: string;
    tokenType?: string;
    expiresIn?: string;
    idToken?: string;
};